Move failed procedures to my labnote
To make the entire workflow easier to read I started documenting failures to
labnote not to project directory. This commit moves previous failures to labnote
so that everything documented here are valuable. I've also indicated in this document
where on my labnote the failures are documented so that it is always possible to go back
to my labnote to read them.